Refactoring to internal/service
