+++ internal/command/semver/semver_test.go
@@ -9,7 +9,7 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	"github.com/moorara/gelato/internal/command"
-	"github.com/moorara/gelato/internal/git"
+	"github.com/moorara/gelato/internal/service/git"
 	"github.com/moorara/gelato/pkg/semver"
 )
