+++ internal/service/decorate/decorate.go
@@ -15,7 +15,7 @@ import (
 
 	"golang.org/x/tools/imports"
 
-	"github.com/moorara/gelato/internal/decorate/modifier"
+	"github.com/moorara/gelato/internal/service/decorate/modifier"
 	"github.com/moorara/gelato/internal/log"
 )
