+++ internal/service/decorate/modifier/generic.go
