+++ internal/service/decorate/test/horizontal/go.mod
