+++ internal/service/decorate/test/invalid/go.mod
