+++ internal/service/decorate/test/vertical/go.mod
